create components folder and dumpads componenet.
